Settings work + FAQ + About Us + Welcome page. Need to implement vibration and sound control.

+++ ScanVentureProject/src/org/tensorflow/demo/Account.java
@@ -4,14 +4,14 @@ import java.io.Serializable;
 
 @SuppressWarnings("serial")
 public class Account implements Serializable {
-    private int level;
-    private String username , id;
+    private long level;
+    private String username , id ;
 
     public String getId() {
         return id;
     }
 
-    public int getLevel() {
+    public long getLevel() {
         return level;
     }
 
@@ -21,7 +21,7 @@ public class Account implements Serializable {
     public Account(String username){
         this.username = username;
     }
-    public Account(String id, String username, int level) {
+    public Account(String id, String username, long level) {
         this.id = id;
         this.username = username;
         this.level = level;
